+++ pet-app/app/login/page.js
@@ -3,20 +3,17 @@ import React, { useEffect, useState } from 'react'
 import MainButton from '../ui/Buttons/MainButton/MainButton'
 import { StyledHeading, StyledLogin, StyledPage, StyledTextField } from './styled';
 import { FormControl, TextField } from '@mui/material';
-import Header from "../components/Header/Header";
+import Header from "../../components/Header/Header";
 import SecondaryButton from '../ui/Buttons/SecondaryButton/SecondaryButton';
-import { Title } from '@mui/icons-material';
 
 export default function Login() {
   const [theme, setTheme] = useState('light'); // Pode ser 'light' ou 'dark'
-
   const prefersDarkMode = window.matchMedia('(prefers-color-scheme: dark)').matches;
   useEffect(() => {
     setTheme(prefersDarkMode ? 'dark' : 'light');
   }, [prefersDarkMode]);
-
   return (
-    <StyledPage>
+    <StyledPage >
       <Header />
       <StyledLogin theme={theme}>
         <section>
@@ -28,8 +25,6 @@ export default function Login() {
             placeholder='seu@email.com' type="email" label={"Email:"} focused required />
           <TextField type="password" label={"Senha:"} required />
           <MainButton title="Login" type="submit" />
-          {/* <br/> */}
-
           <SecondaryButton title="Ainda não tem conta? Clique aqui!" />
         </FormControl >
       </StyledLogin>
