+++ pet-app/app/page.js
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import styles from './page.module.css'
-import Header from './components/Header/Header'
+import Header from '../components/Header/Header'
 
 export default function Home() {
   return (
