+++ pet-app/app/ui/Buttons/MainButton/MainButton.js
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react'
-import { StyledMainButton } from './layout'
+
 import { Button } from '@mui/material';
 
 export default function MainButton({ title, action }) {
