+++ pet-app/components/Header/styled.js
